+++ Sources/NSFetchRequest+Utils.swift
@@ -2,8 +2,6 @@
 
 import CoreData
 
-// TODO: review these methods (probably it's better to add some NSPredicate utils)
-
 extension NSFetchRequest {
   /// Creates a NSFetchRequest.
   ///
@@ -18,28 +16,6 @@ extension NSFetchRequest {
     self.entity = entity
     self.predicate = predicate
     self.fetchBatchSize = batchSize
-  }
-
-  /// - Parameter predicate: A NSPredicate object.
-  /// Associates to `self` a `new` compound NSPredicate formed by **AND**-ing the current predicate with a given `predicate`.
-  @objc
-  public func andPredicate(_ predicate: NSPredicate) {
-    guard let currentPredicate = self.predicate else {
-      self.predicate = predicate
-      return
-    }
-    self.predicate = NSCompoundPredicate(andPredicateWithSubpredicates: [currentPredicate, predicate])
-  }
-
-  /// - Parameter predicate: A NSPredicate object.
-  /// Associates to `self` a `new` compound NSPredicate formed by **OR**-ing the current predicate with a given `predicate`.
-  @objc
-  public func orPredicate(_ predicate: NSPredicate) {
-    guard let currentPredicate = self.predicate else {
-      self.predicate = predicate
-      return
-    }
-    self.predicate = NSCompoundPredicate(orPredicateWithSubpredicates: [currentPredicate, predicate])
   }
 
   /// - Parameter descriptors: An array of NSSortDescriptor objects.